changes add by daniela gavrila
